apps/data/requirements.txt
requests
pyjwt
minio
jsonschema
clerk-backend-api
schedule
selenium
python-dotenv
